extract bucketize to utils file
